new trap simulation
made a new folder for the elevator trap with its own scripts for simulations to not overstep into the current trap simulation.
added in a few models (both stl and f3d formats)
fixed a small deprecation warning for pyvista in the electrostatics.py module
fixed an import error in utils/expansion.py where the program tried to import "matrix" from scipy which does not, and according to google has not ever existed.