generalization of plot_muls and bug fix
generalized the function plot_mul so that it now takes the used_multipole list from the MultipoleControl class. If you tried to take out a multipole before, you got an error.

Also fixed a bug in the III_Multipole_Expansion where running the cell where the MultipoleControl instance is defined without rerunning the entire script would cause the controlled_electrodes variable to grow each time which messed with the solutions. Fixed it by reinitializing the variable at the start of that cell.